backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
chromadb==0.4.17
sentence-transformers==2.2.2
requests==2.31.0
PyGithub==1.59.1
python-multipart==0.0.6
mcp==1.0.0
anthropic-mcp-client==0.1.0